+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fix playlist dragging
+
 ## 0.20.0 - 2025 Oct 4
 - Add audio visualiser (thanks to help from @zachwinter)
 - Add filtering for specific fields in search, like `artist:apashe` or `album:"mirai sekai"`. You can also right-click a column header to add a filter. Filtering by artist also finds featured artists, remixers etc in the song title.